+++ .github/workflows/ci.yml
@@ -60,6 +60,9 @@ jobs:
       #   run: |
       #     cargo tarpaulin --features=wgpu --coveralls=${{ secrets.COVERALLS_REPO_TOKEN }}
 
+      - name: Docs
+        run: cargo doc --no-deps
+
       - name: Build and Test
         run: cargo test --features=flex
