+++ crates/bunsen/src/kits/bimm/mod.rs
@@ -1,0 +1,6 @@
+//! # `bimm` - Bunsen/Burn Image Models
+//!
+//! This is inspired by the `timm` / (Torch Image Models) library.
+
+pub mod resnet;
+pub mod swin;
