+++ crates/bunsen/src/kits/bimm/resnet/downsample.rs
