+++ crates/bunsen/src/kits/bimm/resnet/layer_block.rs
@@ -80,14 +80,14 @@ pub struct LayerBlockContractConfig {
     #[config(default = "None")]
     pub bottleneck_policy: Option<BottleneckPolicyConfig>,
 
-    /// [`crate::compat::normalization_wrapper::Normalization`] config.
+    /// Normalization config.
     ///
     /// The feature size of this config will be replaced
     /// with the appropriate feature size for the input layer.
     #[config(default = "NormalizationConfig::Batch(BatchNormConfig::new(0))")]
     pub normalization: NormalizationConfig,
 
-    /// [`crate::compat::activation_wrapper::Activation`] config.
+    /// Activation config.
     #[config(default = "ActivationConfig::Relu")]
     pub activation: ActivationConfig,
 }
@@ -431,7 +431,7 @@ mod tests {
     use super::*;
     use crate::{
         contracts::assert_shape_contract,
-        kits::imgs::resnet::BasicBlockConfig,
+        kits::bimm::resnet::BasicBlockConfig,
         support::testing::PerfTestBackend,
     };
