+++ crates/bunsen/src/kits/bimm/resnet/residual_block.rs
@@ -35,7 +35,7 @@ use burn::{
 
 use crate::{
     blocks::images::drop::drop_block::DropBlockOptions,
-    kits::imgs::resnet::{
+    kits::bimm::resnet::{
         BasicBlock,
         BasicBlockConfig,
         BasicBlockMeta,
@@ -73,14 +73,14 @@ pub struct ResidualBlockContractConfig {
     #[config(default = "None")]
     pub bottleneck_policy: Option<BottleneckPolicyConfig>,
 
-    /// [`crate::compat::normalization_wrapper::Normalization`] config.
+    /// Normalization config.
     ///
     /// The feature size of this config will be replaced
     /// with the appropriate feature size for the input layer.
     #[config(default = "NormalizationConfig::Batch(BatchNormConfig::new(0))")]
     pub normalization: NormalizationConfig,
 
-    /// [`crate::compat::activation_wrapper::Activation`] config.
+    /// Activation config.
     #[config(default = "ActivationConfig::Relu")]
     pub activation: ActivationConfig,
 }
