+++ crates/bunsen/src/kits/bimm/resnet/stems.rs
