+++ crates/bunsen/src/kits/bimm/swin/v2/swin_model.rs
@@ -46,7 +46,7 @@ use crate::{
         assert_shape_contract_periodically,
         unpack_shape_contract,
     },
-    kits::imgs::swin::v2::{
+    kits::bimm::swin::v2::{
         block_sequence::{
             StochasticDepthTransformerBlockSequence,
             StochasticDepthTransformerBlockSequenceConfig,
