+++ crates/bunsen/src/kits/bimm/swin/v2/window_attention/pos_grid.rs
