+++ crates/bunsen/src/kits/bimm/swin/v2/windowing.rs
