--- crates/bunsen/src/kits/imgs/mod.rs
@@ -1,4 +1,0 @@
-//! # Various Image Model Kits
-
-pub mod resnet;
-pub mod swin;
