+++ crates/bunsen/src/kits/mod.rs
@@ -2,6 +2,6 @@
 //!
 //! This module contains domain-specific complete kits.
 
+pub mod bimm;
 pub mod gpts;
-pub mod imgs;
 pub mod sims;
