+++ demos/bimm/.run/Docs.run.xml
@@ -1,7 +1,7 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="Docs" type="CargoCommandRunConfiguration" factoryName="Cargo Command" focusToolWindowBeforeRun="true">
     <option name="buildProfileId" value="dev" />
-    <option name="command" value="doc --open" />
+    <option name="command" value="doc --no-deps --all-features --open" />
     <option name="workingDirectory" value="file://$PROJECT_DIR$" />
     <envs>
       <env name="CINIC10_PATH" value="/media/Data/CINIC-10" />
